+++ src/stores/nodeStore.ts
@@ -15,7 +15,7 @@ import GeocodingService from '~/services/geocodingService';
 import { IStopAreaItem } from '~/services/stopAreaService';
 import GeometryUndoStore from '~/stores/geometryUndoStore';
 import NodeLocationType from '~/types/NodeLocationType';
-import { roundLatLng, roundLatLngs } from '~/utils/geomHelpers';
+import { roundLatLng, roundLatLngs } from '~/utils/geomUtils';
 import FormValidator from '~/validation/FormValidator';
 import NavigationStore from './navigationStore';
 import ValidationStore, { ICustomValidatorMap } from './validationStore';
