+++ src/utils/geomUtils.ts
